+++ backend/plantios/urls.py
@@ -1,6 +1,12 @@
-from django.urls import path
-from .views import PlantioListCreateView
+from django.urls import include, path
+from rest_framework.routers import DefaultRouter
+
+from .views import PlantioViewSet
+
+router = DefaultRouter()
+
+router.register(r'plantios', PlantioViewSet, basename='plantio')
 
 urlpatterns = [
-	path('plantios/', PlantioListCreateView.as_view(), name='plantio-list-create'),
+    path('', include(router.urls)),
 ]
