+++ backend/plantios/views.py
@@ -1,9 +1,11 @@
-from rest_framework import generics
+from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated
+
 from .models import Plantio
 from .serializers import PlantioSerializer
 
-class PlantioListCreateView(generics.ListCreateAPIView):
+
+class PlantioViewSet(viewsets.ModelViewSet):
     serializer_class = PlantioSerializer
     permission_classes = [IsAuthenticated]
