+++ frontend/src/navigation/AppRoutes.js
@@ -4,21 +4,18 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 
 import AgricultorSolicitacoes from "../pages/Credito/AgricultorSolicitacoesScreen";
 import AnaliseSolicitacoes from "../pages/Credito/AnaliseSolicitacoesScreen";
-import CotacoesScreen from "../pages/CotacoesScreen";
+import CotacoesScreen from "../pages/Cotacoes/CotacoesScreen";
 import EscolhaPropriedadeScreen from "../pages/Previsao/EscolhaPropriedade";
 import WeatherScreen from "../pages/Previsao/PrevisaoScreen";
-import PerguntasScreen from "../pages/PerguntasScreen";
-import RegistrarPlantio from "../pages/RegistrarPlantio";
-import RegistrarPropriedade from "../pages/RegistrarPropriedade";
+import PerguntasScreen from "../pages/Perguntas/PerguntasScreen";
+import RegistrarPlantio from "../pages/Plantio/RegistrarPlantio";
+import RegistrarPropriedade from "../pages/Propriedade/RegistrarPropriedade";
 import PedirCredito from "../pages/Credito/SolicitarCreditoScreen";
 import Home from "../pages/Homes";
 import EscolhaCulturaScreen from "../pages/Calendario/EscolhaCulturaScreen";
 import CalendarioScreen from "../pages/Calendario/CalendarioScreen";
-import ProdutividadeScreen from "../pages/ProdutividadeScreen"
-import AnaliseSolicitacoes from "../pages/AnaliseSolicitacoesScreen";
-import AgricultorSolicitacoes from "../pages/AgricultorSolicitacoesScreen";
-import ListarPlantiosScreen from "../pages/ListarPlantiosScreen";
-
+import ProdutividadeScreen from "../pages/Produtividade/ProdutividadeScreen";
+import ListarPlantiosScreen from "../pages/Plantio/ListarPlantiosScreen";
 
 const Stack = createNativeStackNavigator();
 
@@ -28,9 +25,8 @@ export default function Routes() {
       <Stack.Screen name="Home" component={Home} options={{ headerShown: false }} />
       <Stack.Screen name="Registrar Propriedades" component={RegistrarPropriedade} />
       <Stack.Screen name="Registrar Plantios" component={RegistrarPlantio} />
-      <Stack.Screen name="Produtividade" component={ProdutividadeScreen} />
+      <Stack.Screen name="Relatório de Produtividade" component={ProdutividadeScreen} />
       <Stack.Screen name="Plantios Registrados" component={ListarPlantiosScreen} />
-      {/* <Stack.Screen name="Produção Atual" component={...} /> */}
       <Stack.Screen name="Solicitar Crédito" component={PedirCredito} />
       <Stack.Screen name="Solicitações de Crédito" component={AgricultorSolicitacoes} />
       <Stack.Screen
@@ -53,7 +49,13 @@ export default function Routes() {
       />
       <Stack.Screen name="Previsão do Tempo" component={EscolhaPropriedadeScreen} />
       <Stack.Screen name="Previsões" component={WeatherScreen} />
-      <Stack.Screen name="Calendário de Plantios" component={EscolhaCulturaScreen} />
+      <Stack.Screen
+        name="Calendário de Plantios"
+        component={EscolhaCulturaScreen}
+        options={{
+          headerRight: () => <Text style={styles.subHeader}>CONAB</Text>,
+        }}
+      />
       <Stack.Screen name="Calendário" component={CalendarioScreen} />
     </Stack.Navigator>
   );
