+++ frontend/src/pages/Autenticacao/api.js
@@ -1,0 +1,27 @@
+import { API_URL } from "@env";
+
+const API_BASE = `${API_URL}/api`;
+
+export async function registerUser(data) {
+  const response = await fetch(`${API_BASE}/users/register/`, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(data),
+  });
+
+  const json = await response.json();
+  console.log("Erro ou sucesso no registro:", json);
+  return json;
+}
+
+export async function loginUser(data) {
+  const response = await fetch(`${API_BASE}/users/login/`, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(data),
+  });
+
+  const json = await response.json();
+  console.log("Resultado do login:", json);
+  return json;
+}
